requirements.txt
flask==1.1.2
cryptography==3.4.6
flask-htpasswd==0.4.0
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vis==0.4.1
line-bot-sdk==2.0.1
matplotlib==3.3.3
numpy==1.19.4
opencv-contrib-python==4.5.3.56
Pillow==8.0.1
requests==2.25.0
scikit-image==0.17.2
scipy==1.1.0
